store/requirements.txt
Django==5.0.2
pillow==10.2.0
django-allauth==0.61.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
flake8==7.0.0
isort==5.13.2
stripe==8.2.0
redis==5.0.1
celery==5.3.6
django-redis==5.4.0
psycopg2-binary==2.9.9

